requirements-dev.txt
asgiref==3.5.2
Django==4.1.4
django-cors-headers==3.5.0
django-environ==0.9.0
djangorestframework==3.12.2
Pillow==8.1.2
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
dj-database-url
dj-static==0.0.6
python-decouple==3.3